# justdoit-frontend/src/app/services/todo-api/swagger.json
{"swagger":"2.0","info":{"description":"TODO API reference for developers","title":"TODO API"},"host":"192.168.0.14:9002","tags":[{"name":"test-controller","description":"Test Controller"},{"name":"todo-controller","description":"Todo Controller"},{"name":"user-controller","description":"User Controller"},{"name":"user-relationship-controller","description":"User Relationship Controller"}],"paths":{"/backend/api/relationship/all":{"get":{"tags":["user-relationship-controller"],"summary":"getAllPeople","operationId":"getAllPeopleUsingGET","produces":["*/*"],"parameters":[{"name":"email","in":"header","description":"email","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/UserRelationshipTransport"}}}}}},"/backend/api/relationship/friends":{"get":{"tags":["user-relationship-controller"],"summary":"getAllFriends","operationId":"getAllFriendsUsingGET","produces":["*/*"],"parameters":[{"name":"userId","in":"header","description":"userId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/UserRelationshipTransport"}}}}}},"/backend/api/relationship/pending":{"get":{"tags":["user-relationship-controller"],"summary":"geAllPendingRequests","operationId":"geAllPendingRequestsUsingGET","produces":["*/*"],"parameters":[{"name":"userId","in":"header","description":"userId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/UserRelationshipTransport"}}}}}},"/backend/api/relationship/requested":{"get":{"tags":["user-relationship-controller"],"summary":"getAllRequestedRequests","operationId":"getAllRequestedRequestsUsingGET","produces":["*/*"],"parameters":[{"name":"userId","in":"header","description":"userId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/UserRelationshipTransport"}}}}}},"/backend/api/relationship/{id}":{"get":{"tags":["user-relationship-controller"],"summary":"get","operationId":"getUsingGET_1","produces":["*/*"],"parameters":[{"name":"email","in":"header","description":"email","required":true,"type":"string"},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"},{"name":"userId","in":"header","description":"userId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserRelationshipTransport"}}}},"post":{"tags":["user-relationship-controller"],"summary":"create","operationId":"createUsingPOST_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"email","in":"header","description":"email","required":true,"type":"string"},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserRelationshipTransport"}}}},"delete":{"tags":["user-relationship-controller"],"summary":"delete","operationId":"deleteUsingDELETE_1","produces":["*/*"],"parameters":[{"name":"email","in":"header","description":"email","required":true,"type":"string"},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"204":{"description":"No Content"}}}},"/backend/api/relationship/{id}/accept":{"put":{"tags":["user-relationship-controller"],"summary":"accept","operationId":"acceptUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"email","in":"header","description":"email","required":true,"type":"string"},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"},{"name":"userId","in":"header","description":"userId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserRelationshipTransport"}}}}},"/backend/api/relationship/{id}/cancel":{"delete":{"tags":["user-relationship-controller"],"summary":"cancel","operationId":"cancelUsingDELETE","produces":["*/*"],"parameters":[{"name":"email","in":"header","description":"email","required":true,"type":"string"},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"204":{"description":"No Content"}}}},"/backend/api/relationship/{id}/reject":{"put":{"tags":["user-relationship-controller"],"summary":"reject","operationId":"rejectUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"email","in":"header","description":"email","required":true,"type":"string"},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"},{"name":"userId","in":"header","description":"userId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserRelationshipTransport"}}}}},"/backend/api/test":{"get":{"tags":["test-controller"],"summary":"validateResetToken","operationId":"validateResetTokenUsingGET","produces":["*/*"],"parameters":[{"name":"userId","in":"header","description":"userId","required":true,"type":"string"}],"responses":{"204":{"description":"No Content"}}}},"/backend/api/todos":{"post":{"tags":["todo-controller"],"summary":"create","operationId":"createUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"email","in":"header","description":"email","required":true,"type":"string"},{"in":"body","name":"todoTransport","description":"todoTransport","required":true,"schema":{"$ref":"#/definitions/TodoTransport"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TodoTransport"}}}}},"/backend/api/todos/assigned":{"get":{"tags":["todo-controller"],"summary":"getAssignedTodos","operationId":"getAssignedTodosUsingGET","produces":["*/*"],"parameters":[{"name":"email","in":"header","description":"email","required":true,"type":"string"},{"name":"oldestFirst","in":"query","description":"oldestFirst","required":false,"type":"boolean","default":true}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/TodoTransport"}}}}}},"/backend/api/todos/created":{"get":{"tags":["todo-controller"],"summary":"getCreatedTodos","operationId":"getCreatedTodosUsingGET","produces":["*/*"],"parameters":[{"name":"email","in":"header","description":"email","required":true,"type":"string"},{"name":"oldestFirst","in":"query","description":"oldestFirst","required":false,"type":"boolean","default":true}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/TodoTransport"}}}}}},"/backend/api/todos/{todoId}":{"get":{"tags":["todo-controller"],"summary":"get","operationId":"getUsingGET","produces":["*/*"],"parameters":[{"name":"todoId","in":"path","description":"todoId","required":true,"type":"integer","format":"int64"},{"name":"userId","in":"header","description":"userId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TodoTransport"}}}},"put":{"tags":["todo-controller"],"summary":"update","operationId":"updateUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"todoId","in":"path","description":"todoId","required":true,"type":"integer","format":"int64"},{"in":"body","name":"todoTransport","description":"todoTransport","required":true,"schema":{"$ref":"#/definitions/TodoTransport"}},{"name":"userId","in":"header","description":"userId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TodoTransport"}}}},"delete":{"tags":["todo-controller"],"summary":"delete","operationId":"deleteUsingDELETE","produces":["*/*"],"parameters":[{"name":"todoId","in":"path","description":"todoId","required":true,"type":"integer","format":"int64"},{"name":"userId","in":"header","description":"userId","required":true,"type":"string"}],"responses":{"204":{"description":"No Content"}}}},"/backend/api/todos/{todoId}/finished":{"put":{"tags":["todo-controller"],"summary":"markFinished","operationId":"markFinishedUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"todoId","in":"path","description":"todoId","required":true,"type":"integer","format":"int64"},{"name":"userId","in":"header","description":"userId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TodoTransport"}}}}},"/backend/api/todos/{todoId}/not-finished":{"put":{"tags":["todo-controller"],"summary":"markNotFinished","operationId":"markNotFinishedUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"todoId","in":"path","description":"todoId","required":true,"type":"integer","format":"int64"},{"name":"userId","in":"header","description":"userId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TodoTransport"}}}}},"/backend/users/create":{"post":{"tags":["user-controller"],"summary":"createUser","operationId":"createUserUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"email","in":"query","description":"email","required":true,"type":"string"},{"name":"id","in":"query","description":"id","required":true,"type":"integer","format":"int64"},{"name":"username","in":"query","description":"username","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserTransport"}}}}}},"securityDefinitions":{"Bearer %token":{"type":"apiKey","name":"Authorization","in":"header"}},"definitions":{"TodoTransport":{"type":"object","required":["assignee","description","dueDate","title"],"properties":{"assignee":{"$ref":"#/definitions/UserTransport"},"createdAt":{"type":"string","format":"date-time"},"createdBy":{"$ref":"#/definitions/UserTransport"},"description":{"type":"string"},"dueDate":{"type":"string","format":"date-time"},"finishedAt":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"status":{"type":"string","enum":["CREATED","EXPIRED","FINISHED"]},"title":{"type":"string"},"updatedAt":{"type":"string","format":"date-time"}},"title":"TodoTransport"},"UserRelationshipTransport":{"type":"object","properties":{"relatedUser":{"$ref":"#/definitions/UserTransport"},"requestedAt":{"type":"string","format":"date-time"},"userRelationshipType":{"type":"string","enum":["BLOCKED","FRIENDS","NONE","PENDING"]}},"title":"UserRelationshipTransport"},"UserTransport":{"type":"object","properties":{"email":{"type":"string"},"id":{"type":"integer","format":"int64"},"username":{"type":"string"}},"title":"UserTransport"}}}